moved all header components into labbook header
