+++ src/js/components/dashboard/labbooks/Labbooks.js
@@ -9,7 +9,7 @@ import WizardModal from 'Components/wizard/WizardModal'
 import Loader from 'Components/shared/Loader'
 import LocalLabbooksContainer, {LocalLabbooks} from 'Components/dashboard/labbooks/localLabbooks/LocalLabbooks'
 import RemoteLabbooks from 'Components/dashboard/labbooks/remoteLabbooks/RemoteLabbooks'
-import LoginPrompt from 'Components/labbook/branchMenu/LoginPrompt'
+import LoginPrompt from 'Components/labbook/labbookHeader/branchMenu/LoginPrompt'
 import ToolTip from 'Components/shared/ToolTip'
 //utils
 import Validation from 'JS/utils/Validation'
