+++ src/js/components/labbook/Labbook.js
@@ -1,6 +1,6 @@
 //vendor
 import React, { Component } from 'react'
-import { Route , Switch, Link} from 'react-router-dom';
+import { Route , Switch} from 'react-router-dom';
 import {
   createFragmentContainer,
   graphql
@@ -12,7 +12,7 @@ import { connect } from 'react-redux'
 //store
 import store from "JS/redux/store"
 //components
-import ContainerStatus from './containerStatus/ContainerStatus'
+import LabbookHeader from './labbookHeader/LabbookHeader'
 import Login from 'Components/login/Login';
 import Activity from './activity/Activity'
 import Code from './code/Code'
@@ -21,9 +21,6 @@ import OutputData from './outputData/OutputData'
 import Overview from './overview/Overview'
 import Environment from './environment/Environment'
 import Loader from 'Components/shared/Loader'
-import Branches from './branches/Branches'
-import BranchMenu from './branchMenu/BranchMenu'
-import ToolTip from 'Components/shared/ToolTip';
 import ErrorBoundary from 'Components/shared/ErrorBoundary'
 //utils
 import {getFilesFromDragEvent} from "JS/utils/html-dir-content";
@@ -37,10 +34,7 @@ class Labbook extends Component {
     localStorage.setItem('owner', store.getState().routes.owner)
 
     //bind functions here
-    this._setSelectedComponent = this._setSelectedComponent.bind(this)
     this._setBuildingState = this._setBuildingState.bind(this)
-    this._showLabbookModal = this._showLabbookModal.bind(this)
-    this._hideLabbookModal = this._hideLabbookModal.bind(this)
     this._toggleBranchesView = this._toggleBranchesView.bind(this)
     this._branchViewClickedOff = this._branchViewClickedOff.bind(this)
 
@@ -66,16 +60,20 @@ class Labbook extends Component {
       }
     })
   }
+
   /**
     @param {}
     subscribe to store to update state
     set unsubcribe for store
   */
   componentDidMount() {
+
     this._setStickHeader();
+
     window.addEventListener('scroll', this._setStickHeader)
     window.addEventListener('click', this._branchViewClickedOff )
   }
+
   /**
     @param {}
     removes event listeners
@@ -90,8 +88,8 @@ class Labbook extends Component {
     window.removeEventListener('scroll', this._setStickHeader)
 
     window.removeEventListener('click', this._branchViewClickedOff)
-
   }
+
   /**
     @param {event}
     updates state of labbook when prompted ot by the store
@@ -100,6 +98,23 @@ class Labbook extends Component {
   _branchViewClickedOff(evt) {
     if(evt.target.className.indexOf('Labbook__veil') > -1) {
       this._toggleBranchesView(false, false);
+    }
+  }
+
+  /**
+    @param {string}
+    makes branch name pretty
+    @return {string}
+  */
+  _sanitizeBranchName(branchName){
+    const username = localStorage.getItem('username')
+    const workspace = `gm.workspace-${username}`
+
+    if(branchName){
+
+      const prettyBranchName = (branchName === workspace) ? 'workspace' : branchName.replace(`${workspace}.`, '')
+
+      return prettyBranchName
     }
   }
 
@@ -132,24 +147,7 @@ class Labbook extends Component {
       })
     }
   }
-  /**
-    @param {string} componentName - input string componenetName
-    updates state of selectedComponent
-    updates history prop
-  */
-  _setSelectedComponent = (componentName) =>{
-    if(componentName !== this.props.selectedComponent){
-      if(store.getState().detailView.selectedComponent === true){
-        store.dispatch({
-          type: 'UPDATE_DETAIL_VIEW',
-          payload: {
-            detailMode: false
-          }
-        })
 
-      }
-    }
-  }
   /**
     @param {boolean} isBuilding
     updates container status state
@@ -164,147 +162,6 @@ class Labbook extends Component {
         payload:{
           'isBuilding': isBuilding
         }
-      })
-    }
-  }
-
-  /**
-  @param {boolean} isSyncing
-  updates container status state
-  updates labbook state
-*/
-  _setSyncingState = (isSyncing) => {
-    this.refs['ContainerStatus'] && this.refs['ContainerStatus'].setState({ 'isSyncing': isSyncing })
-
-    if (this.props.isSyncing !== isSyncing) {
-      store.dispatch(
-        {
-          type: 'IS_SYNCING',
-          payload: {
-            'isSyncing': isSyncing
-          }
-        })
-    }
-  }
-
-  /**
-    @param {boolean} isPublishing
-    updates container status state
-    updates labbook state
-  */
- _setPublishingState = (isPublishing) => {
-
-    this.refs['ContainerStatus'] && this.refs['ContainerStatus'].setState({ 'isPublishing': isPublishing })
-
-    if (this.props.isPublishing !== isPublishing) {
-      store.dispatch(
-        {
-          type: 'IS_PUBLISHING',
-          payload: {
-            'isPublishing': isPublishing
-          }
-        })
-    }
-  }
-
-  /**
-    @param {boolean} isExporting
-    updates container status state
-    updates labbook state
-  */
-  _setExportingState = (isExporting) => {
-
-    this.refs['ContainerStatus'] && this.refs['ContainerStatus'].setState({ 'isExporting': isExporting })
-
-    if (this.props.isExporting !== isExporting) {
-      store.dispatch(
-        {
-          type: 'IS_EXPORTING',
-          payload: {
-            'isExporting': isExporting
-          }
-        })
-    }
-  }
-
-
-  /**
-    @param {object} item
-    returns nav jsx
-  */
-
-  _changeSlider() {
-    let pathArray = this.props.location.pathname.split('/')
-    let selectedPath = (pathArray.length > 4 ) ? pathArray[pathArray.length - 1] : 'overview'
-    let defaultOrder = ['overview', 'activity', 'environment', 'code', 'inputData', 'outputData'];
-    let selectedIndex = defaultOrder.indexOf(selectedPath);
-    return (
-      <hr className={' Labbook__navigation-slider Labbook__navigation-slider--' + selectedIndex}/>
-    )
-  }
-
-  _getNavItem(item, index){
-    let pathArray = this.props.location.pathname.split('/')
-    let selectedPath = (pathArray.length > 4 ) ? pathArray[pathArray.length - 1] : 'overview' // sets avtive nav item to overview if there is no menu item in the url
-    let liCSS = classNames({
-      'selected': selectedPath === item.id,
-      ['Labbook__navigation-item Labbook__navigation-item--' + item.id]: !selectedPath !== item.id,
-      ['Labbook__navigation-item--' + index]: true,
-    });
-    return (
-      <li
-        id={item.id}
-        key={item.id}
-        className={liCSS}
-        onClick={()=> this._setSelectedComponent(item.id)}
-        >
-        <Link
-          onClick={this._scrollToTop}
-          to={`../../../projects/${this.props.owner}/${this.props.match.params.labbookName}/${item.id}`}
-          replace
-        >
-          {item.name}
-        </Link>
-      </li>
-    )
-  }
-  /**
-    @param {}
-    updates html element classlist and labbook state
-  */
-  _showLabbookModal = () => {
-    if(!this.props.modalVisible){
-      store.dispatch(
-        {
-          type: 'MODAL_VISIBLE',
-          payload:{
-            'modalVisible': true
-          }
-      })
-    }
-
-  }
-  /**
-    @param {}
-    updates html element classlist and labbook state
-  */
-  _hideLabbookModal = () => {
-
-    if(document.getElementById('labbookModal')){
-      document.getElementById('labbookModal').classList.add('hidden')
-    }
-
-    if(document.getElementById('modal__cover')){
-      document.getElementById('modal__cover').classList.add('hidden')
-    }
-
-    if(this.props.modalVisible){
-      store.dispatch(
-        {
-          type: 'MODAL_VISIBLE',
-          payload:{
-            'modalVisible': false
-          }
       })
     }
   }
@@ -338,20 +195,6 @@ class Labbook extends Component {
     }
   }
   /**
-    @param {string}
-    makes branch name pretty
-    @return {string}
-  */
-  _sanitizeBranchName(branchName){
-    const username = localStorage.getItem('username')
-    const workspace = `gm.workspace-${username}`
-    if(branchName){
-      const prettyBranchName = (branchName === workspace) ? 'workspace' : branchName.replace(`${workspace}.`, '')
-
-      return prettyBranchName
-    }
-  }
-  /**
     scrolls to top of window
   */
   _scrollToTop(){
@@ -361,15 +204,14 @@ class Labbook extends Component {
   render(){
 
     const { isAuthenticated } = this.props.auth
-    const {labbookName} = this.props
     const isLockedBrowser = {locked: (this.props.isPublishing || this.props.isSyncing || this.props.isExporting), isPublishing: this.props.isPublishing, isExporting: this.props.isExporting, isSyncing: this.props.isSyncing}
     const isLockedEnvironment = this.props.isBuilding || this.props.isSyncing || this.props.isPublishing
 
     if(this.props.labbook){
 
-      const {labbook} = this.props
-      const name = this._sanitizeBranchName(this.props.labbook.activeBranchName)
-      const {branchesOpen} = this.props
+      const {labbook, branchesOpen} = this.props
+      const branchName = this._sanitizeBranchName(this.props.labbook.activeBranchName)
+
       const labbookCSS = classNames({
         'Labbook': true,
         'Labbook--detail-mode': this.props.detailMode,
@@ -377,154 +219,30 @@ class Labbook extends Component {
         'is-demo': window.location.hostname === Config.demoHostName,
       })
 
-      const branchNameCSS = classNames({
-        'Labbook__branch-title': true,
-        'Labbook__branch-title--open': branchesOpen,
-        'Labbook__branch-title--closed': !branchesOpen
-      })
-
-      const labbookHeaderCSS = classNames({
-        'Labbook__header': true,
-        'is-sticky': this.props.isSticky,
-        'is-expanded': this.props.isExpanded
-      })
-
-      const {visibility} = this.props.labbook
-      const labbookLockCSS = classNames({
-        [`Labbook__${visibility}`]: true,
-        [`Labbook__${visibility}--sticky`]: this.props.isSticky
-      })
-
       return(
-        <div
-          className={labbookCSS}>
+        <div className={labbookCSS}>
 
-           <div className="Labbook__inner-container flex flex--row">
-             <div className="Labbook__component-container flex flex--column">
-              <div className="Labbook__header-container">
-                <div className={labbookHeaderCSS}>
-                  <div className="Labbook__row-container">
-                   <div className="Labbook__column-container--flex-1">
-                     <div className="Labbook__name-title">
-                       <div>
-                       {`${labbook.owner}/${labbookName}${this.props.isSticky ? '/ ': ''}`}
-                       </div>
-                       {
-                         this.props.isSticky &&
-                         <div className="Labbook__name-branch">{name}</div>
-                       }
+          <div className="Labbook__inner-container flex flex--row">
 
-                       {  ((visibility === 'private') ||
-                           (visibility === 'public')) &&
+            <div className="Labbook__component-container flex flex--column">
 
-                           <div className={labbookLockCSS}></div>
-                       }
+               <LabbookHeader
+                 setBuildingState={this._setBuildingState}
+                 toggleBranchesView={this._toggleBranchesView}
+                 branchName={branchName}
+                 {...this.props} />
 
-                       {
-                         this.props.isExpanded &&
-                         <div className="Labbook__navigation-container--header flex-0-0-auto column-1-span-11">
-                           <ul className="Labbook__navigation Labbook__navigation--header flex flex--row">
-                            {
-                              Config.navigation_items.map((item, index) => {
-                                return (this._getNavItem(item, index))
-                              })
-                            }
-                            {
-                              this._changeSlider()
-                            }
-                          </ul>
-                        </div>
-                       }
-                     </div>
-
-                     <div className={branchNameCSS}>
-                       <div className="Labbook__name" onClick={()=> this._toggleBranchesView(!branchesOpen, false)}>
-                           {name}
-                       </div>
-                       <div
-                         onClick={()=> this._toggleBranchesView(!branchesOpen, false)}
-                        className="Labbook__branch-toggle"></div>
-                        <ToolTip section="branchView"/>
-                     </div>
-
-                </div>
-                <div className="Labbook__column-container">
-
-                   <BranchMenu
-                     visibility={visibility}
-                     description={labbook.description}
-                     history={this.props.history}
-                     collaborators={labbook.collaborators}
-                     defaultRemote={labbook.defaultRemote}
-                     labbookId={labbook.id}
-                     remoteUrl={labbook.overview.remoteUrl}
-                     setSyncingState={this._setSyncingState}
-                     setPublishingState={this._setPublishingState}
-                     setExportingState={this._setExportingState}
-                     isExporting={this.props.isExporting}
-                     toggleBranchesView={this._toggleBranchesView}
-                     isMainWorkspace={name === 'workspace' || name === `gm.workspace-${localStorage.getItem('username')}`}
-                     auth={this.props.auth}
-                    />
-                    <ErrorBoundary type="containerStatusError" key="containerStatus">
-                      <ContainerStatus
-                        ref="ContainerStatus"
-                        base={labbook.environment.base}
-                        containerStatus={labbook.environment.containerStatus}
-                        imageStatus={labbook.environment.imageStatus}
-                        labbookId={labbook.id}
-                        setBuildingState={this._setBuildingState}
-                        isBuilding={this.props.isBuilding}
-                        isSyncing={this.props.isSyncing}
-                        isPublishing={this.props.isPublishing}
-                        creationDateUtc={labbook.creationDateUtc}
-                      />
-                    </ErrorBoundary>
-                  </div>
-                </div>
-                <div className={(this.props.branchesOpen) ? "Labbook__branches-container":" Labbook__branches-container Labbook__branches-container--collapsed"}>
-
-                  <div className={(this.props.branchesOpen) ? 'Labbook__branches-shadow Labbook__branches-shadow--upper' : 'hidden'}></div>
-                  <ErrorBoundary type={this.props.branchesOpen ?  'branchesError': 'hidden'} key="branches">
-                    <Branches
-                      defaultRemote={labbook.defaultRemote}
-                      branchesOpen={this.props.branchesOpen}
-                      labbook={labbook}
-                      labbookId={labbook.id}
-                      activeBranch={labbook.activeBranchName}
-                      toggleBranchesView={this._toggleBranchesView}
-                      mergeFilter={this.props.mergeFilter}
-                      setBuildingState={this._setBuildingState}
-                    />
-                  </ErrorBoundary>
-
-                  <div className={(this.props.branchesOpen) ? 'Labbook__branches-shadow Labbook__branches-shadow--lower' : 'hidden'}></div>
-                </div>
-              </div>
-                <div className="Labbook__navigation-container mui-container flex-0-0-auto">
-                   <ul className="Labbook__navigation flex flex--row">
-                     {
-                       Config.navigation_items.map((item, index) => {
-                         return (this._getNavItem(item, index))
-                       })
-                     }
-                     {
-                      (this._changeSlider())
-                     }
-                   </ul>
-                 </div>
-
-             </div>
-
-             <div className="Labbook__view mui-container flex flex-1-0-auto">
+               <div className="Labbook__view mui-container flex flex-1-0-auto">
 
                   <Switch>
                     <Route
                       exact
                       path={`${this.props.match.path}`}
                       render={() => {
+
                         return (
                           <ErrorBoundary type="labbookSectionError">
+
                             <Overview
                               key={this.props.labbookName + '_overview'}
                               labbook={labbook}
@@ -536,18 +254,26 @@ class Labbook extends Component {
                               isPublishing={this.props.isPublishing}
                               scrollToTop={this._scrollToTop}
                             />
+
                           </ErrorBoundary>
                         )
                       }}
                     />
 
                     <Route path={`${this.props.match.path}/:labbookMenu`}>
+
                       <Switch>
+
                         <Route
                           path={`${this.props.match.path}/overview`}
                           render={() => {
+
                             return (
-                              <ErrorBoundary type="labbookSectionError" key="overview">
+
+                              <ErrorBoundary
+                                type="labbookSectionError"
+                                key="overview">
+
                                 <Overview
                                   key={this.props.labbookName + '_overview'}
                                   labbook={labbook}
@@ -559,6 +285,7 @@ class Labbook extends Component {
                                   isPublishing={this.props.isPublishing}
                                   scrollToTop={this._scrollToTop}
                                 />
+
                               </ErrorBoundary>
                             )
                           }}
@@ -569,17 +296,21 @@ class Labbook extends Component {
                           render={() => {
 
                           return (
-                            <ErrorBoundary type="labbookSectionError" key="activity">
+                            <ErrorBoundary
+                              type="labbookSectionError"
+                              key="activity">
+
                               <Activity
                                 key={this.props.labbookName + '_activity'}
                                 labbook={labbook}
                                 activityRecords={this.props.activityRecords}
                                 labbookId={labbook.id}
                                 activeBranch={labbook.activeBranch}
-                                isMainWorkspace={name === 'workspace'}
+                                isMainWorkspace={branchName === 'workspace'}
                                 setBuildingState={this._setBuildingState}
                                 {...this.props}
                               />
+
                             </ErrorBoundary>
                           )
                         }} />
@@ -587,8 +318,12 @@ class Labbook extends Component {
                         <Route
                           path={`${this.props.match.url}/environment`}
                           render={() => {
+
                             return (
-                              <ErrorBoundary type="labbookSectionError" key="environment">
+                              <ErrorBoundary
+                                type="labbookSectionError"
+                                key="environment">
+
                                 <Environment
                                   key={this.props.labbookName + '_environment'}
                                   labbook={labbook}
@@ -599,49 +334,72 @@ class Labbook extends Component {
                                   isLocked={isLockedEnvironment}
                                   {...this.props}
                                 />
-                              </ErrorBoundary>
-                              )
+
+                              </ErrorBoundary>)
                           }}
                         />
 
-                        <Route path={`${this.props.match.url}/code`} render={() => {
-                          return (
-                            <ErrorBoundary type="labbookSectionError" key="code">
-                              <Code
-                                labbook={labbook}
-                                labbookId={labbook.id}
-                                setContainerState={this._setContainerState}
-                                isLocked={isLockedBrowser}
-                              />
-                            </ErrorBoundary>
-                            )
+                        <Route
+                          path={`${this.props.match.url}/code`}
+                          render={() => {
+
+                            return (
+                              <ErrorBoundary
+                                type="labbookSectionError"
+                                key="code">
+
+                                  <Code
+                                    labbook={labbook}
+                                    labbookId={labbook.id}
+                                    setContainerState={this._setContainerState}
+                                    isLocked={isLockedBrowser}
+                                  />
+
+                               </ErrorBoundary>)
+                            }}
+                          />
+
+                        <Route
+                          path={`${this.props.match.url}/inputData`}
+                          render={() => {
+
+                            return (
+                              <ErrorBoundary
+                                type="labbookSectionError"
+                                key="input">
+
+                                  <InputData
+                                    labbook={labbook}
+                                    labbookId={labbook.id}
+                                    isLocked={isLockedBrowser}
+                                  />
+
+                             </ErrorBoundary>)
                         }} />
 
-                        <Route path={`${this.props.match.url}/inputData`} render={() => {
-                          return (
-                            <ErrorBoundary type="labbookSectionError" key="input">
-                              <InputData
-                                labbook={labbook}
-                                labbookId={labbook.id}
-                                isLocked={isLockedBrowser}
-                              />
-                            </ErrorBoundary>
-                              )
-                        }} />
+                        <Route
+                          path={`${this.props.match.url}/outputData`}
+                          render={() => {
 
-                        <Route path={`${this.props.match.url}/outputData`} render={() => {
-                          return (
-                            <ErrorBoundary type="labbookSectionError" key="output">
-                              <OutputData
-                                labbook={labbook}
-                                labbookId={labbook.id}
-                                isLocked={isLockedBrowser}
-                              />
-                            </ErrorBoundary>
-                            )
-                        }} />
+                            return (
+                              <ErrorBoundary
+                                type="labbookSectionError"
+                                key="output">
+
+                                      <OutputData
+                                        labbook={labbook}
+                                        labbookId={labbook.id}
+                                        isLocked={isLockedBrowser}
+                                      />
+
+                              </ErrorBoundary>)
+                            }}
+                          />
+
                       </Switch>
+
                     </Route>
+
                   </Switch>
 
               </div>
@@ -649,14 +407,18 @@ class Labbook extends Component {
             </div>
 
           </div>
+
           <div className="Labbook__veil"></div>
-        </div>
-      )
+
+       </div>)
 
     }else{
+
       if(isAuthenticated()){
+
         return (<Loader />)
       }else{
+
         return (<Login auth={this.props.auth}/>)
       }
     }
