+++ src/js/components/labbook/labbookHeader/branchMenu/DeleteLabbook.js
