+++ src/js/components/labbook/labbookHeader/branchMenu/LoginPrompt.js
