+++ src/js/components/labbook/labbookHeader/branches/DeleteBranch.js
